requirements.txt
streamlit
requests
beautifulsoup4
spacy
python-dotenv
google-generativeai
python-docx 
PyPDF2
weasyprint
fpdf
langdetect
easyocr
streamlit-extras
# python -m spacy download en_core_web_sm
# for pymupdf to work properly, you may need to install additional system dependencies:
# sudo apt update
# sudo apt install libpango-1.0-0 libpango1.0-dev

